+++ packages/core/src/index.ts
@@ -9,6 +9,7 @@ export type {
   Interpolator,
   LoopOpts,
   ParallelValues,
+  StaggerArrayOpts,
   StaggerFn,
   StaggerFrom,
   StaggerOpts,